loader: don't read bitcoind's stdout to avoid a deadlock
It was possible to get into a state where we were waiting for lianad to
start, which was waiting on a response from bitcoind which in turn was
waiting on the GUI to make room in its piped stdout buffer. Deadlock.

Instead of getting its logs from stdout, read its debug.log file.

Co-Authored-By: Antoine Poinsot <darosior@protonmail.com>