gui: support user/password RPC authentication
This change updates the GUI to support the latest lianad config
file with the user/password option for bitcoind RPC authentication.

The settings have been updated to set either cookie file path
or user and password.

The installer still only supports the cookie file and has only
been updated with the changes required to compile.

When starting, the internal bitcoind now looks for the cookie file
in the usual location rather than relying on the config file.