Bump MSRV for lianad to 1.63
1.63 is the latest version supported by Debian. It's also now almost a
year and a half old (August 2023). We start getting way behind on our
dependency updates, and sometimes even missing some security updates
(nothing which affects us but still..).

Asking to support a less than 2.5yo compiler shouldn't be that much to
ask, but hey. 1.63 is now a better balance between support for upstream
updates and support for a not recklessly new compiler.