fuzz: fuzzing integrations, fuzz the descriptors module
This integrates fuzzing into our project by introducing two targets
which exercise the descriptor parsing and analysis logic.

The `descriptor_parse` is dead simple but not very effective. The
`descriptors` harness tries to be smarter by almost always generating a
valid Liana descriptor.

Of course, this is just a first integration and both could be made more
effective.