Use tokio mpsc channel sender and receiver
in Tokio, when you call .recv().await on a Receiver, it will
asynchronously wait for a message to be sent to the channel. This means
that the task will yield control back to the Tokio runtime until a
message is available, allowing other tasks to run concurrently.

If the Sender side of the channel is dropped or closed, recv().await
will return None, indicating that no more messages will be sent.