gui(installer): show manually entered xpub in modal until saved
This fixes a regression introduced as part of the key-from-token
work for cosigner and safety net keys.

The manually entered xpub should continue to be displayed until
the "Apply" button has been clicked and the changes saved.

The `chosen_key_source_kind` field has been renamed to
`form_key_source_kind` as it is only used to determine which of
the xpub and token forms should be displayed.