Use type wrapper instead of PathBuf for directories
With the multiplicity of directories, liana directory,
network directory and in the future per wallet lianad
directories, passing a simple type PathBuf does not
transmit to the contributors the information about
which directory each module works with.

This commit introduces wrappers around PathBuf
to give this information and link parent directories
to their child directories.